gesture_recognization/requirements.txt
# Computer Vision và Hand Detection
opencv-python>=4.8.0
mediapipe>=0.10.0

# Thư viện máy học
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration và HTTP
PyYAML>=6.0
requests>=2.31.0
pyserial

# Visualization (Optional)
matplotlib>=3.7.0
seaborn>=0.12.0
cvzone
requests
